Changed Alarm Logic
chkAlarm function was checking for alarms to trigger that occured in the future, i.e next 7 days, 12 hours, 30 minutes. 

I am unsure if this functionality was intended by the original maintainer but it seems weird to me. 

Now the chkAlarm will check if an alarm is meant to trigger *now* or in the past and output to chat if so.

Signed-off-by: Big-T <84488170+BigDashT@users.noreply.github.com>