Added functionality to alarms
Added new functionality to alarms. 

Alarms can now be configured to halt time advancements if the time advancement would put the end time to being past the alarm. If enabled, the alarm will pause time advancement and set the alarm to the time of the alarm.

Also added an option for the alarm text to either be sent to chat for all players to see, or for alarms to be whispered to the GM.

Alarm text is contained in a nice neat box which looks nice. So should draw attention to it if players see it! 

Signed-off-by: Big-T <84488170+BigDashT@users.noreply.github.com>